requirements.txt
numpy>=1.24
scipy>=1.10
scikit-image>=0.21
scikit-learn>=1.3
matplotlib>=3.7
Pillow>=10.0
tifffile>=2023.7
jupyter>=1.0
ipykernel>=6.25
streamlit>=1.28
mahotas>=1.4
torch>=2.1
torchvision>=0.16
pandas>=2.0
